Trying to reduce main: handlers separated
