+++ internal/handlers/get.go
@@ -1,0 +1,82 @@
+package handlers
+
+import (
+	"context"
+	"encoding/json"
+	"errors"
+	"fmt"
+	"net/http"
+	"os"
+	"time"
+
+	"github.com/KirillZiborov/lnkshortener/internal/auth"
+	"github.com/KirillZiborov/lnkshortener/internal/config"
+	"github.com/go-chi/chi"
+	"github.com/jackc/pgx/v5/pgxpool"
+)
+
+func GetHandler(cfg config.Config, store URLStore) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+
+		id := chi.URLParam(r, "id")
+		shortenedURL := fmt.Sprintf("%s/%s", cfg.BaseURL, id)
+
+		originalURL, deleted, err := store.GetOriginalURL(shortenedURL)
+		if err != nil {
+			if errors.Is(err, os.ErrProcessDone) {
+				http.Error(w, "Not found", http.StatusNotFound)
+			} else {
+				http.Error(w, "Internal server error", http.StatusInternalServerError)
+			}
+			return
+		}
+
+		if deleted {
+			http.Error(w, "URL has been deleted", http.StatusGone)
+			return
+		}
+
+		w.Header().Set("Location", originalURL)
+		w.WriteHeader(http.StatusTemporaryRedirect)
+	}
+}
+
+func PingDBHandler(db *pgxpool.Pool) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+
+		ctx, cancel := context.WithTimeout(r.Context(), 1*time.Second)
+		defer cancel()
+
+		err := db.Ping(ctx)
+		if err != nil {
+			http.Error(w, "Unable to connect to database", http.StatusInternalServerError)
+			return
+		}
+
+		w.WriteHeader(http.StatusOK)
+	}
+}
+
+func GetUserURLsHandler(store URLStore) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		userID, err := auth.AuthGet(r)
+		if err != nil {
+			http.Error(w, "Unathorized", http.StatusUnauthorized)
+		}
+
+		records, err := store.GetUserURLs(userID)
+		if err != nil {
+			http.Error(w, "Failed to get a list of user's URLs", http.StatusInternalServerError)
+			return
+		}
+
+		if len(records) == 0 {
+			w.WriteHeader(http.StatusNoContent)
+			return
+		}
+
+		w.Header().Set("Content-Type", "application/json")
+		w.WriteHeader(http.StatusOK)
+		json.NewEncoder(w).Encode(records)
+	}
+}
